requirements.txt
# API
fastapi[standard]
uvicorn
python-dotenv
pydantic
asyncpg
sqlalchemy
psycopg2-binary
python-jose
requests
email-validator

# AI
transformers
datasets
torch --index-url https://download.pytorch.org/whl/cpu
contractions
google-genai